feat: Restrict usage of SYSCON resources
Add a private field to the structs of those resources to restrict their
use. From now on, they must be used via the `System` struct.

The `*const ()` part in the `PhantomData` additionally makes sure that
`Send` and `Sync` are not implemented for those types. I haven't really
thought about whether those should be available, so it may become
necessary to implement them later.