refactor: Implement SYSCON traits for references
Changes the implementations of various SYSCON-related traits,
implementing them for references to peripheral types, instead of the
peripheral types themselves.

I'm about to change some SYSCON APIs so they require a mutable reference
to the resources they are modifying. This commit is a prerequisite for
that, as it allows peripherals that own a `&lpc82x::Peripheral` to pass
a `&mut &lpc82x::Peripheral`.

This commit introduces some complexity in the `usart` module. This
should be resolved, once the recent changes to svd2rust[0] arrive
downstream.

[0] https://github.com/japaric/svd2rust/pull/158