feat: Restructure `syscon` module
Split `SYSCON` into `Api`, which fills the role that `SYSCON` previously
filled, and `SYSCON`, a struct that only contains `Api`. This doesn't
make much sense by itself, but it is a first step towards restructuring
everything.

I'll not go into my whole vision here in this commit message, but as far
as the `syscon` module is concerned, all its singleton structs that
currently live within `System` will be moved to `SYSCON`. I plan to do
the same for other peripheral modules.