feat: Prevent changes to `GPIO` while pin exists
Add a shared reference from each pin back to `GPIO`, to prevent it from
being changed in any way that could affect the operation of the pin.

As an intended side effect of this change, it is no longer possible to
create instances of pin types from outside the crate. This is a first
step towards solving the various safety issues of the pin API.