feat: Add type `Pin`
This type is used to represent pins. The previously used types, like
`PIO0_0`, still exist, but are simplified and serve as a type parameter
for `Pin`, to identify the specific pin.

The reason for this change is the growing complexity of the pin types.
They all had a field, a set of methods, and in the future they would
have required additional type parameter to represent the pin state. I
think it's nicer to concentrate this within a single type, instead of
implementing it for all pin types.