refactor: Create functions in single macro call
Changes the macro to create fixed functions to generate all fixed
functions in a single macro invocation, instead of one invocation per
fixed function. This makes it possible to extend the macro to also
generate a struct that contains all fixed functions.

This commit lays the groundwork for changing the API for fixed functions
in the same way as the API for movable functions was changed, making
fixed functions more prominent and enabling them to be extended with
type state later.