feat: Prevent user from creating movable functions
This change prevents the user from creating the movable function types,
by adding a private field to them.

The private field is a reference to the register that the movable
function controls a part of. This means that those types have everything
they need and are no longer dependent on the internal structure of
`swm::Api`.