feat: Prevent user from creating fixed functions
This change prevents the user from creating the fixed function types, by
adding a private field to them.

The private field is a reference to the register that the fixed function
controls a bit in. This means that those types have everything they need
and are no longer dependent on the internal structure of `swm::Api`.