feat: Restrict `FixedFunction` methods
Restrict access to the methods of `FixedFunction` by requiring the pin
type that is owned by `Pin`. This change is a prerequisite for the
upcoming safety improvements of the pin API.

As an intended side effect, it is no longer possible to use the methods
of one `Pin` instance to enable or disable fixed functions for another
pin instance.