refactor: Don't keep reference to SWM in functions
Remove the reference to the SWM registers from the types that represent
fixed and movable functions. While all of them only access
non-overlapping bits, those bits can't be written to atomically.

The only reason this didn't cause race conditions previously, was
because the methods writing to those bits required a mutable reference
to `SWM` (which is also required for other reasons). If the mutable
reference is going to be required anyway, we might as well make the code
a bit more robust against future modifications and access the registers
that way.