feat: Require user to disable pin functions
Require the user to disable or unassign pin functions manually before
making the pin available for GPIO. The way this was done previously was
no good, in my opinion, as it had a runtime cost and only took care of
fixed functions.

I have an idea for how to improve the current situation and assist the
user in disabling or unassigning any functions, but we'll have to see
how that works out. For now, requiring the user to do it manually is
good enough, I think.