feat: Track movable function state
Add a type parameter to track the current state of a movable function.
This change solves one half of the safety puzzle for movable functions:
It is no longer possible to assign a movable function to multiple pins.
However, it is still possible to assign multiple conflicting movable
functions to the same pin.