feat: Restrict methods to input functions
Restrict the non-output function assignment methods to input functions.
Previously, they would work with all functions, which was a leftover
from the time when there was no distinction between input and output
functions (that sounds like it was a long time since then, but it's
actually been just an hour or so).

This change prevents users from erroneously using the non-output
function to (un)assign output functions, which would work on the
hardware level, but not update the type state of the pin type.