refactor: Store mutable refs to raw peripherals
The peripheral APIs get a mutable reference to the raw peripheral
structs, but then store only a shared reference. This makes no
difference to the public API, as the lifetimes are what guarantee the
exclusive access, but it seems weird.

This change is also a prerequisite for a simplification to the SYSCON
module that I'm currently working on.