refactor: Add traits to control state transitions
Add traits to control SWM-related pin state transitions and implement
those for `Pin`. As you may notice, the types in there are redundant
with the types in the `impl` blocks currently used to control pin state
transitions. These new traits will allow for another reduction of the
SWM API surface, which I'll realize in follow-up commits.