chore: Enable compiletest on Travis
Because of a bug in compiletest[0], it needs to run while the target
directory is somewhat clean. To achieve that, I've disabled caching of
Cargo artifacts. I don't think this is a big sacrifice, as I have the
suspicion that downloading and updating the cache takes more time than
just rebuilding everything.

[0] https://github.com/laumann/compiletest-rs/issues/114