feat: Make splitting PMU API an explicit step
This is an initial step towards a radical simplification of the API
state management. I'm going to re-introduce the `Peripherals` struct,
which will give me a known initial state, which will allow for the
aforementioned radical simplification.

However, it will also mean that users won't have access to raw
peripherals before the HAL API has been initialized. To give users
access to the raw peripherals where necessary, we'll need a method that
destroys the HAL peripheral and returns the raw peripheral. This will no
longer be possible after the raw peripheral has been split, hence this
change.