fix: Add compiler fences around DMA transfer
The compiler is free to reorder reads and writes as it sees fit. This
can lead to erroneous code, as the compiler has no understanding of what
the DMA controller does, and therefore might move a write to the buffer
into the transfer.

These compiler fences should prevent the compiler from doing any such
thing.