Re-add workspace
It was previously removed, as it caused problems when building the
crates[1] and we assumed that it didn't add much value.

I've since discovered that it is still valuable to have it, for running
things like `cargo update` and `cargo outdated` on all crates at the
same time. This is especially valuable for me, as I have a local script
that goes through a number of projects I maintain to figure out which
ones need attention.

I think it's somewhat problematic that just running `cargo build` won't
lead to the desired results, which could be confusing, but I believe
that the benefits of fixing my local automation outweigh this.

[1] https://github.com/lpc-rs/lpc8xx-hal/pull/154