Remove default target selection
I've come to the conclusion that having a default target selection is
more confusing that not having one. Not selecting a target results in a
clear error message, while selecting a target while forgetting to
unselect the default one could be confusing to new users.

"You have to select a target" is also easier to explain in documentation
than the default target, as I'm currently finding out.

Given this, I believe the default target is not worth it, especially
since most users will select a target in their `Cargo.toml` anyway.