Remove core peripherals from `Peripherals`
Re-export `pac::CorePeripherals` as a convenient alternative.

The core peripherals don't really need to be part of `Peripherals`.
We're not actually doing anything with them; it's just slightly more
convenient for the user. Although not much convenience is lost, as the
pretty minimal changes to the examples show.

The reason I'm removing them is to improve interoperability with RTFM,
as that assumes control over the core peripherals. As of before this
commit, using this HAL with RTFM would give the user safe access to two
sets of core peripherals, which seems like asking for trouble.