Improve split between USART's `Rx` and `Tx`
It was already possible to split a USART peripheral into a receiver and
transmitter, but that feature was mostly useless. Both still borrowed
the original USART, which severely limited the use cases. It wasn't
possible to move a receiver or transmitter somewhere else, without also
moving the USART or moving the USART into a `static`.

This commit changes this, by creating a strong split that doesn't
require any borrowing. `Rx` and `Tx` can just be moved out of `USART`
and used independently. To do anything with `USART` that would affect
either of them (disabling it, for example), they have to be moved back
in.

The `RegProxy` stuff got a bit crazy here, but I wasn't able to find a
simpler solution. The only alternative I can think of is not use
`RegProxy` and instead unsafely access the peripheral everywhere it is
used.