Make dependency on Termion optional
It doesn't build on Windows, which I think is enough reason to make it
optional and unselected by default.

I don't think having it as an optional dependency is particularly
useful, but I didn't want to remove it outright. I'm hoping that
Termion's problems on Windows are temporary, and once it is fixed,
restoring the previous functionality will be easy.

Close #212