Make public peripheral constructors `pub(crate)`
Peripheral constructors in this HAL were originally never public, to
allow the HAL APIs to uphold the guarantees they make. (Upholding any
API guarantees expressed via type state is not possible when
constructing the API from a PAC peripheral, since the state of that is
not known.) That they are public now is an artifact from when we had
multiple crates for portability.

This commit restores the previous state, thereby making the API less
confusing and error-prone. Users that need to construct HAL API from PAC
peripherals for some reason can still do so using unsafe means.