Don't require `GpioPin` to borrow GPIO peripheral
This makes `GpioPin` much more flexible, allowing it to be moved to
other contexts easily, while preserving the API guarantees.

Previously, the reference to `GPIO` was used to prevent anyone from
disabling the GPIO peripheral while a pin was used for GPIO. This
guarantee is now upheld by the token being moved out of `GPIO`, which
makes it impossible to move to another state.