Un-derive `Copy` from `gpio::Registers`
It doesn't make much of a difference, but it's not necessary, and having
it be `Copy` for no reason seems icky (although I can't point to any
concrete situation where it could lead to problems right now).

Honestly, I don't even know why I made it `Copy` in the first place.
Probably a temporary brain fart.