Set initial level when switching to output mode
Previously, there was no way to control the initial output level of a
pin, when switching it to output mode. This doesn't matter for a
blinking LED, but there are situations where an undesired signal is not
acceptable.

This commit rectifies this, by requiring the initial level to be passed
as an argument. This is a bit more tedious for the common case, but more
flexible, and overall less error-prone (I've had bugs in the past, that
were caused by me not thinking of the initial output level).

Close #53