Remove `Instance` trait bounds from USART structs
I actually don't like removing the trait bound, as I think it makes the
code less clear. It doesn't make that much of a difference, however, and
there's a good reason for removing it: `const fn` methods that involve
trait bounds (not necessarily on themselves, but on the `impl` blocks
where they are defined) are unstable at this point.

This means that implementing a `const fn` method on a struct that
includes a trait bound is probably impossible. This isn't an issue in
this crate (at the moment), but it affects any structs that contain
`USART`, `Rx`, or `Tx`.

With this change, implementing `const fn` methods on such structs
becomes possible, as long as those methods don't require any trait
bounds to work.