Split `swm` module into multiple files
The module is huge, and finding things became pretty hard. This adds a
bit of overhead in terms of lines of code, but I think it's worth it.

This is only a first stab. I'm sure there are better ways to organize
this, but I think this is already better than what we had previously.