Add PININT API
The API is pretty basic so far, but it's enough to be useful. In any
case, I don't need any more features right now.

The name is a bit difficult here, as the peripheral has different names:

- The chapter in the user manual only refers to it as "Pin interrupts/
  pattern match engine", without providing a short name, like most other
  peripherals have.
- The SVD file calls it PINT, which seems unnecessarily short.
- The field names in the SYSCON registers refer to its clock as
  GPIO_INT, which is not accurate (the user manual hints that the pin
  can be in any state, not just GPIO).

I decided to adapt the name of the peripheral's interrupts, as specified
in the user manual. It seems reasonable short and descriptive, which
matching the other peripheral names well enough.