Remove `-tui` option from GDB command
Recent versions of GDB treat missing source code for a file as a fatal
error, if `-tui` is enabled. This is not good, as missing source code is
quite normal, for example if your MCU is currently executing code from
the core library. Older versions just didn't show any code in that case.

I regularly run into the case that I can't upload a program using `cargo
run`, because the MCU happens to be executing something from the core
library (or at least GDB thinks it is). The only reliable way to get out
of this seems to be to remove the `-tui`.

I found that manually adding `-tui` whenever you need it is way more
convenient than having to deal with this issue all the time, hence this
change.