Reduce number of target selection `cfg`s
This creates one module per target in `usart::clock` and moves all
target-specific code there. I believe this is a bit cleaner, as only one
`cfg` per target is required.

The disadvantage are the redundant imports, but I believe it's a
worthwhile trade-off. Especially since the imports required for the
different targets are going to diverge going forward, which would
require even more `cfg`s.