Consolidate target-specific clock code
The `&()` I had to add isn't pretty, but I think it's worth it. We can
define more targeted types in the future, to make this more
understandable. And to fix (insert words to prevent GitHub from closing
that issue) #205, we might have to anyway.

In any case, this is how far I'm willing to go right now, and I think
it's good enough.