Remove default type state from `I2C`
I'm working on adding support for I2C slave mode. A type parameter to
identify which mode the peripheral is in will be added to `Enabled` (as
opposed to `I2C`; the mode is only important if the peripheral is
enabled, and anything else would make implementing methods that are
always available, if the peripheral is enabled, more complex than need
be).

I think with this new mode parameter in the mix, it is no longer
obviously better to have a default value for the state parameter. With
default type parameters, there's always a trade-off between brevity and
convenience on one side, and clarity on the other. While master mode
will certainly be more common, I think the brevity/convenience gained by
using it as the default will not be worth the loss of clarity.

This is not a clear-cut decision and I don't feel strongly about it. I'm
open to changing this and making enabled/master the new default.