Add infallible GPIO API
I'm currently writing an introductory article to embedded Rust using this
crate, and all the `unwrap`s all over the code stand out as really gross
and hard to explain. I think, in principle, the case is pretty clear:
The operations are infallible, and we should expose that in the API.
That we also implement a more general interface that has to take
fallibility into account is an orthogonal concern.

However, things get a bit muddled in practice. The methods on the trait
and the inherent methods have the same name, and this could be very
confusing. However, in the latest `embedded-hal` master (to be released
as version 1.0, I believe) those fallible methods have been renamed with
a `try_` prefix. This makes any confusion introduced by this commit
temporary.

We could wait until `embedded-hal` 1.0 has been released and make this
change then. However, as I'm writing that article, I'd like to get this
change in now. Since that article might expose LPC8xx HAL to a
potentially large audience (it's going to be published in a German print
magazine for software developers), I believe the confusion and
complexity saved in the article outweighs the temporary confusion caused
by this commit.