Add internal method that waits until I2C is ready
Note that this isn't just a refactoring, as the loops this replaces
weren't all identical. One of them didn't check for errors, and some of
them had a wait condition that was a bit sketchy (they read MSTSTATE
without checking MSTPENDING, even though MSTSTATE only reflects the
state when MSTPENDING is set, according to the documentation).