Revamp I2C API
The I2C API supports enabling master and slave mode independently, and
operate both at the same time. I initially didn't plan to expose this
capability for now, due to the added complexity.

Turns out I need it for testing now (I'm having yet unresolved problems
with the test setup I was planning to use) so here we go. This commit
doesn't add support for slave mode yet, but it prepares the rest of the
API for this.