Fix slave mode not working independently of master
The I2C peripheral on LPC845 can be clocked from various sources, and in
addition to enabling the peripheral clock, this source needs to be
chosen. I had assumed this was only required for master mode, as it is
related to SCL clock generation. This isn't made clear in the user
manual, as far as I can tell.

As it turns out, this is necessary for slave mode too. It will only work
partially without the clock being selected. The specific problem I was
seeing, was that after software told the peripheral to ACK an address,
the peripheral didn't do so, and instead kept stretching the clock.

I'm not crazy about the added complexity (especially the `PhantomData`
in the type signature irks me), but I don't see a better way to do it
right now.