Re-export `void` crate
We expose it in type signatures, which means our users may need to
depend on it. I believe it is best practive to re-export it in this
case, as that relieves the user from having to manage their own
dependency on it (while keeping the version in sync at all times).