Don't require user to pass DMA descriptor table
This makes using the DMA peripheral more convenient and opens the door
for further simplications, at the cost of requiring more memory when the
DMA peripheral isn't used (actually I believe that as of this commit,
the compiler can optimize the static out, if DMA is not used; I'm
working on further changes that would make this impossible).

I think the added memory use if far from ideal, but right now I'd like
to focus on a simpler and more flexible API. We can restore efficiency
later, if desired, by being a bit more clever about using the static and
allowing the compiler to optimize it.