Enable/disable DMA peripheral as a whole
This removes the need for carrying a reference in every enabled channel,
thereby removing the lifetime from channels and transfers, while
offering the same guarantees. This makes the API much more convenient in
non-trivial use cases.

In addition, it allows for disabling the peripheral after a channel has
been disabled, which has not been possible with the old API.