Panic, if length of DMA buffer is zero
The previous approach of just returning an unconfigured transfer seemed
a bit sketchy, and I haven't tested it at all in my recent work (it's
been there since the initial version of the DMA API).

I think these belong in the same category as the other buffer length
restrictions, and those currently panic. Maybe the method should return
a `Result` instead and they all should be error variants, but I think
for now the panics are fine.