Add bespoke `Enabled` type for USART
The general `init_state::Enabled` struct is flexible, but it can lead to
some unpleasantness, as seen with the whole clock configuration stuff in
the I2C API, that involves `PhantomData` in the type signature.

I am about to make some changes that would introduce the same kind of
ugliness in the USART API. Adding a bespoke `Enabled` type will make
this a bit nicer.