Generate `Flag` and `Interrupts` types for USART
This is an improved version the macros founds in other peripheral
modules. It can be used to not only enable/disable interrupts with a
minimal amount of repetition and manual code, it allows the same for
querying/resetting flags.

I believe it should be relatively straight-forward to make this macro
peripheral-agnostic and move it to a shared module, where it can be used
for all of the peripheral APIs.

As a final note, I think this macro is right at the edge of complexity
where a procedural macro starts to make sense. I don't want to go into
that right now, as that would require adding an additional crate, but I
think we should keep that in mind as an option for the future.