Add support for hardware-based address detection
This adds a critical section to the code which can't be controlled by
the user. We've discussed this before and basically agree that it would
be nicer to give the user the option to pass a critical section token.

We haven't come up with a good pattern for that though, and since the
hardcoded critical section seems good enough for now, I decided to go
with this simple but non-optimal solution again.