Add inherent methods to MRT channels
This is a step in preparation for implementing the `CountDown` trait
from the alpha version of embedded-hal. These have different names
(`try_` prefix) and are fallible. Since we will most likely drop support
for the old version of embedded-hal at some point, I thought it best to
add the infallible versions as inherent methods now, instead of having
to make another change later.