Implement `embedded_hal::Pwm` for `CTIMER`
This is ugly. Much uglier than I expected it to be.

I don't know if this is any better than having to check runtime errors
(from a user's perspective it probably is, especially if efficiency is a
concern), but since the old embedded-hal traits don't support error
checking, I don't know if we have much of a choice right now.