add GenericPin
as a drop-in substitute for the `PIO0_0`, `PIO0_1`... etc. structs which are generated at compile
time and used to bind each pin to a `pins::Token`.
`GenericPin`s are *only* intended for use cases which are made impossible by the default
`pins::Trait` having pin information compiled into its type, e.g. if you'd like to store
several pin instances in a collection, or modify them in batches.