Upgrade to latest RTIC alpha
The build was broken due to the release of the new RTIC alpha version.
We pin a specific RTIC version in Cargo.toml to prevent just that, but
that won't prevent Cargo from happily picking up breaking changes from
transitive dependencies. Since RTIC consists of multiple crates, that is
what happened here.

We could explicitly specify and pin all those transitive dependencies in
Cargo.toml to prevent that from happening, but that would get unwieldy.
I decided that it's probably best to just upgrade to the latest alpha,
ride it out for this alpha cycle, and be more careful about adding alpha
dependencies in the future.